Fix bug in game restart and extract player icon color into SCSS
This happened because Gamefield.init is overridden by child classes.
Because of this, when this.init is called in Gamefield.constructor it
doesn't call Gamefield.init but instead calls FieldOfGamefields.init or
SingleGamefield.init

Moved the SVG stroke color of both player icons into the SCSS file.

Player icons are now grey when the FieldOfGamefields is inactive

Did some more refactoring (mostly renaming)